patch zone serial to use int64's instead of int32
According to RFC1035 section 3.3.13 the serial number of a SOA resource
record is a 32 bit unsigned integer. There is no current way of
specifying unsigned integers in the swagger spec, therefore let's patch
it with int64's for now while we figure out what we want to do within
the swagger spec itself.